Move labelBot logic into pytorchBot
Delete `error on issue` test, as it is already covered by basing
pytorchBot tests
