+++ torchci/lib/bot/pytorchBot.ts
@@ -1,5 +1,5 @@
 import { Probot } from "probot";
-import { addComment, reactOnComment } from "./botUtils";
+import { addComment, addLabels, reactOnComment } from "./botUtils";
 import { getHelp, getParser, getInputArgs } from "./cliParser";
 import shlex from "shlex";
 
@@ -100,6 +100,43 @@ function pytorchBot(app: Probot): void {
         );
       }
     }
+
+    async function existingRepoLabels(): Promise<string[]> {
+      const labels = await ctx.octokit.paginate(
+        "GET /repos/{owner}/{repo}/labels",
+        {
+          owner: owner,
+          repo: repo,
+        }
+      );
+      return labels.map((d: any) => d.name);
+    }
+
+    async function handleLabel(labels: string[]) {
+      /**
+       * 1. Get all existing repo labels
+       * 2. Parse labels from command
+       * 3. Find valid and invalid labels
+       * 4. Add valid labels to pr, report invalid labels
+       */
+      const repoLabels = new Set(await existingRepoLabels());
+      // remove unnecessary spaces from labels
+      const labelsToAdd = labels.map((s: string) => s.trim());
+
+      const filteredLabels = labelsToAdd.filter((l: string) => repoLabels.has(l));
+      const invalidLabels = labelsToAdd.filter((l: string) => !repoLabels.has(l));
+      if (invalidLabels.length > 0) {
+        await addComment(
+          ctx,
+          "Didn't find following labels among repository labels: " +
+            invalidLabels.join(",")
+        );
+      }
+      if (filteredLabels.length > 0) {
+        await addLabels(ctx, filteredLabels);
+        await reactOnComment(ctx, "+1");
+    }
+    }
     const skipUsers = [
       54816060, // pytorch-bot
       97764156, // pytorchmergebot
@@ -154,6 +191,9 @@ function pytorchBot(app: Probot): void {
           args.branch = "viable/strict";
         }
         return await handleRebase(args.branch);
+      }
+      case "label": {
+        return await handleLabel(args.labels);
       }
       default:
         return await handleConfused();
